Install jest, capitalize testing complete
This commit marks the initialization of this repo,
beginning with the installation of npm and jest,
creation of .gitignore, package.json, and the JS
files for tests and the functions to be written.
The first tests written are for a function which
takes a string and returns a new string with the
first letter capitalized and the rest turned to
lower case. The tests written for this function
have passed.